requirements.txt
chess
python-dotenv
ipython
numpy
pandas
scikit-learn
stockfish
tensorflow
